+++ tests/test_ionization.py
@@ -22,7 +22,7 @@ through external fields.
 # -------
 # Imports
 # -------
-import math
+import os, shutil, math
 import numpy as np
 from scipy.constants import c, m_e, m_p, e
 # Import the relevant structures in FBPIC
@@ -30,13 +30,14 @@ from fbpic.main import Simulation, adapt_to_grid
 from fbpic.particles import Particles
 from fbpic.lpa_utils.external_fields import ExternalField
 from fbpic.lpa_utils.boosted_frame import BoostConverter
-from fbpic.openpmd_diag import FieldDiagnostic, ParticleDiagnostic
+from fbpic.openpmd_diag import ParticleDiagnostic, BoostedParticleDiagnostic
+# Import openPMD-viewer for checking output files
+from opmd_viewer import OpenPMDTimeSeries
 
 # ----------
 # Parameters
 # ----------
 use_cuda = True
-output_hdf5_files = False
 
 def run_simulation( gamma_boost ):
     """
@@ -49,8 +50,8 @@ def run_simulation( gamma_boost ):
         The Lorentz factor of the frame in which the simulation is carried out.
     """
     # The simulation box
-    zmax = 20.e-6    # Length of the box along z (meters)
-    zmin = 0.e-6
+    zmax_lab = 20.e-6    # Length of the box along z (meters)
+    zmin_lab = 0.e-6
     Nr = 3           # Number of gridpoints along r
     rmax = 10.e-6    # Length of the box along r (meters)
     Nm = 2           # Number of modes used
@@ -70,7 +71,7 @@ def run_simulation( gamma_boost ):
     boost = BoostConverter(gamma_boost)
     # Boost the different quantities
     beta_boost = np.sqrt( 1. - 1./gamma_boost**2 )
-    zmin, zmax = boost.static_length( [zmin, zmax] )
+    zmin, zmax = boost.static_length( [zmin_lab, zmax_lab] )
     p_zmin, p_zmax = boost.static_length( [p_zmin, p_zmax] )
     n_e, = boost.static_density( [n_e] )
     # Increase the number of particles per cell in order to keep sufficient
@@ -80,9 +81,9 @@ def run_simulation( gamma_boost ):
 
     # The laser
     a0 = 1.8         # Laser amplitude
-    lambda0 = 0.8e-6 # Laser wavelength
+    lambda0_lab = 0.8e-6 # Laser wavelength
     # Boost the laser wavelength before calculating the laser amplitude
-    lambda0, = boost.copropag_length( [ lambda0 ], beta_object=1. )
+    lambda0, = boost.copropag_length( [ lambda0_lab ], beta_object=1. )
     # Duration and initial position of the laser
     ctau = 10.*lambda0
     z0 = -2*ctau
@@ -108,7 +109,10 @@ def run_simulation( gamma_boost ):
     v_plasma, = boost.velocity( [ 0. ] )
 
     # The diagnostics
-    diag_period = int(N_step/50) # Period of the diagnostics in number of timesteps
+    diag_period = N_step-1 # Period of the diagnostics in number of timesteps
+
+    # Move into directory `tests`
+    os.chdir('./tests')
 
     # Initialize the simulation object
     sim = Simulation( Nz, zmax, Nr, rmax, Nm, dt,
@@ -139,11 +143,14 @@ def run_simulation( gamma_boost ):
         ExternalField( laser_func, 'By', B0, 0. ) ]
 
     # Add a field diagnostic
-    if output_hdf5_files:
-        sim.diags = [ FieldDiagnostic( diag_period, sim.fld, sim.comm ),
-                        ParticleDiagnostic( diag_period,
-                        {"electrons":sim.ptcl[0], "ions":sim.ptcl[1]},
-                        comm=sim.comm) ]
+    sim.diags = [ ParticleDiagnostic( diag_period,
+                    {"ions":sim.ptcl[1]}, comm=sim.comm) ]
+    if gamma_boost > 1:
+        T_sim_lab = (2.*40.*lambda0_lab + zmax_lab-zmin_lab)/c
+        sim.diags.append( BoostedParticleDiagnostic(zmin_lab, zmax_lab, v_lab=0.,
+            dt_snapshots_lab=T_sim_lab/2., Ntot_snapshots_lab=3,
+            gamma_boost=gamma_boost, period=diag_period, fldobject=sim.fld,
+            species={"ions": sim.ptcl[1]}, comm=sim.comm) )
 
     # Run the simulation
     sim.step( N_step, use_true_rho=True )
@@ -159,6 +166,31 @@ def run_simulation( gamma_boost ):
     N5_fraction = n_N5 / ntot
     print('N5+ fraction: %.4f' %N5_fraction)
     assert ((N5_fraction > 0.30) and (N5_fraction < 0.34))
+
+    # Check consistency in the regular openPMD diagnostics
+    ts = OpenPMDTimeSeries('./diags/hdf5/')
+    last_iteration = ts.iterations[-1]
+    w, q = ts.get_particle( ['w', 'charge'], species="ions",
+                                iteration=last_iteration )
+    # Check that the openPMD file contains the same number of N5+ ions
+    n_N5_openpmd = np.sum(w[ (4.5*e < q) & (q < 5.5*e) ])
+    assert np.isclose( n_N5_openpmd, n_N5 )
+    # Remove openPMD files
+    shutil.rmtree('./diags/')
+
+    # Check consistency of the back-transformed openPMD diagnostics
+    if gamma_boost > 1.:
+        ts = OpenPMDTimeSeries('./lab_diags/hdf5/')
+        last_iteration = ts.iterations[-1]
+        w, q = ts.get_particle( ['w', 'charge'], species="ions",
+                                iteration=last_iteration )
+        # Check that the openPMD file contains the same number of N5+ ions
+        n_N5_openpmd = np.sum(w[ (4.5*e < q) & (q < 5.5*e) ])
+        assert np.isclose( n_N5_openpmd, n_N5 )
+        # Remove openPMD files
+        shutil.rmtree('./lab_diags/')
+
+    os.chdir('../')
 
 def test_ionization_labframe():
     run_simulation(1.)